jobreach_backend/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
python-decouple==3.8
pandas==2.1.3
openai==1.52.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0
